fix(activecampaign): add required sdate param (#1940)
This PR fixes an issue where the ActiveCampaign campaign_status endpoint now has an additional required param that we weren't sending. As a result we were getting a 500 response which prevented newsletters from being sent when AC was the ESP. This PR also surfaces any future errors like this in our logs.